refactor: rename `offline` to `local`
