+++ internal/local/check.go
@@ -1,4 +1,4 @@
-package offline
+package local
 
 import (
 	"fmt"
