+++ internal/local/zip_test.go
@@ -1,4 +1,4 @@
-package offline_test
+package local_test
 
 import (
 	"archive/zip"
@@ -16,7 +16,7 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/google/osv-scanner/internal/offline"
+	"github.com/google/osv-scanner/internal/local"
 	"github.com/google/osv-scanner/pkg/models"
 )
 
@@ -157,10 +157,10 @@ func TestNewZippedDB_Offline_WithoutCache(t *testing.T) {
 	})
 	defer cleanupTestServer()
 
-	_, err := offline.NewZippedDB(testDir, "my-db", ts.URL, true)
+	_, err := local.NewZippedDB(testDir, "my-db", ts.URL, true)
 
-	if !errors.Is(err, offline.ErrOfflineDatabaseNotFound) {
-		t.Errorf("expected \"%v\" error but got \"%v\"", offline.ErrOfflineDatabaseNotFound, err)
+	if !errors.Is(err, local.ErrOfflineDatabaseNotFound) {
+		t.Errorf("expected \"%v\" error but got \"%v\"", local.ErrOfflineDatabaseNotFound, err)
 	}
 }
 
@@ -191,7 +191,7 @@ func TestNewZippedDB_Offline_WithCache(t *testing.T) {
 		"GHSA-5.json": {ID: "GHSA-5"},
 	}))
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, true)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, true)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -211,7 +211,7 @@ func TestNewZippedDB_BadZip(t *testing.T) {
 	})
 	defer cleanupTestServer()
 
-	_, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	_, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err == nil {
 		t.Errorf("expected an error but did not get one")
@@ -224,7 +224,7 @@ func TestNewZippedDB_UnsupportedProtocol(t *testing.T) {
 	testDir, cleanupTestDir := createTestDir(t)
 	defer cleanupTestDir()
 
-	_, err := offline.NewZippedDB(testDir, "my-db", "file://hello-world", false)
+	_, err := local.NewZippedDB(testDir, "my-db", "file://hello-world", false)
 
 	if err == nil {
 		t.Errorf("expected an error but did not get one")
@@ -256,7 +256,7 @@ func TestNewZippedDB_Online_WithoutCache(t *testing.T) {
 	})
 	defer cleanupTestServer()
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -290,7 +290,7 @@ func TestNewZippedDB_Online_WithoutCacheAndNoHashHeader(t *testing.T) {
 	})
 	defer cleanupTestServer()
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -330,7 +330,7 @@ func TestNewZippedDB_Online_WithSameCache(t *testing.T) {
 
 	cacheWrite(t, determineStoredAtPath(testDir, "my-db"), cache)
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -370,7 +370,7 @@ func TestNewZippedDB_Online_WithDifferentCache(t *testing.T) {
 		"GHSA-3.json": {ID: "GHSA-3"},
 	}))
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -402,7 +402,7 @@ func TestNewZippedDB_Online_WithCacheButNoHashHeader(t *testing.T) {
 		"GHSA-3.json": {ID: "GHSA-3"},
 	}))
 
-	_, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	_, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err == nil {
 		t.Errorf("expected an error but did not get one")
@@ -432,7 +432,7 @@ func TestNewZippedDB_Online_WithBadCache(t *testing.T) {
 
 	cacheWriteBad(t, determineStoredAtPath(testDir, "my-db"), "this is not json!")
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
@@ -460,7 +460,7 @@ func TestNewZippedDB_FileChecks(t *testing.T) {
 	})
 	defer cleanupTestServer()
 
-	db, err := offline.NewZippedDB(testDir, "my-db", ts.URL, false)
+	db, err := local.NewZippedDB(testDir, "my-db", ts.URL, false)
 
 	if err != nil {
 		t.Fatalf("unexpected error \"%v\"", err)
